gui: use component::list::see_more in receive view
